# A lot of optimization (234% faster!)