--- a/src/mercury_engine_data_structures/formats/bres.py
+++ b/src/mercury_engine_data_structures/formats/bres.py
@@ -3,10 +3,8 @@ import construct
 from mercury_engine_data_structures.formats import BaseResource, standard_format
 from mercury_engine_data_structures.game_check import Game
 
-BRES = standard_format.game_model('CEnvironmentSoundPresets', 0x02020001)
-
 
 class Bres(BaseResource):
     @classmethod
     def construct_class(cls, target_game: Game) -> construct.Construct:
-        return BRES
+        return standard_format.game_model('CEnvironmentSoundPresets', 0x02020001)