--- a/src/mercury_engine_data_structures/formats/brfld.py
+++ b/src/mercury_engine_data_structures/formats/brfld.py
@@ -8,14 +8,13 @@ from mercury_engine_data_structures.formats import BaseResource, standard_format
 from mercury_engine_data_structures.game_check import Game
 
 logger = logging.getLogger(__name__)
-_BRFLD = standard_format.game_model('CScenario', 0x02000031)
 
 
 class Brfld(BaseResource):
     @classmethod
     @functools.lru_cache
     def construct_class(cls, target_game: Game) -> construct.Construct:
-        return _BRFLD.compile()
+        return standard_format.game_model('CScenario', 0x02000031)
 
     def actors_for_layer(self, name: str) -> dict:
         return self.raw.Root.pScenario.rEntitiesLayer.dctSublayers[name].dctActors