--- a/src/mercury_engine_data_structures/formats/brsa.py
+++ b/src/mercury_engine_data_structures/formats/brsa.py
@@ -6,14 +6,12 @@ from construct import Construct, Container
 from mercury_engine_data_structures.formats import BaseResource, standard_format
 from mercury_engine_data_structures.game_check import Game
 
-_BRSA = standard_format.game_model('CSubAreaManager', 0x02010002)
-
 
 class Brsa(BaseResource):
     @classmethod
     @functools.lru_cache
     def construct_class(cls, target_game: Game) -> Construct:
-        return _BRSA.compile()
+        return standard_format.game_model('CSubAreaManager', 0x02010002)
 
     @property
     def subarea_setups(self) -> Iterator[Container]: