BEFORE:
starter preset: 10.6s
april fools: 13.2s

AFTER:
starter preset: 3.169s
april fools: 5.46s


Door Lock rando has a very slow O(n^3) path filtering which can be fixed in so many ways :P